Refine compute test env
