+++ cpp/src/arrow/compute/kernels/vector_cumulative_ops_test.cc
@@ -37,10 +37,6 @@
 namespace arrow {
 namespace compute {
 
-// Initialize the compute module
-::testing::Environment* compute_kernels_env =
-    ::testing::AddGlobalTestEnvironment(new ComputeKernelEnvironment);
-
 static const std::vector<std::string> kCumulativeFunctionNames{
     "cumulative_sum",          "cumulative_sum_checked", "cumulative_prod",
     "cumulative_prod_checked", "cumulative_min",         "cumulative_max",
