+++ cpp/src/arrow/compute/kernels/vector_sort_test.cc
@@ -46,10 +46,6 @@ using internal::checked_pointer_cast;
 
 namespace compute {
 
-// Initialize the compute module
-::testing::Environment* compute_kernels_env =
-    ::testing::AddGlobalTestEnvironment(new ComputeKernelEnvironment);
-
 #ifdef ARROW_VALGRIND
 using RealArrowTypes = ::testing::Types<FloatType>;
