+++ cpp/src/arrow/dataset/discovery_test.cc
@@ -23,7 +23,6 @@
 #include <memory>
 #include <utility>
 
-#include "arrow/compute/kernels/test_util_internal.h"
 #include "arrow/dataset/partition.h"
 #include "arrow/dataset/test_util_internal.h"
 #include "arrow/filesystem/test_util.h"
@@ -35,11 +34,6 @@ using testing::SizeIs;
 
 namespace arrow {
 
-using compute::ComputeKernelEnvironment;
-
-// Initialize the compute module
-::testing::Environment* compute_kernels_env =
-    ::testing::AddGlobalTestEnvironment(new ComputeKernelEnvironment);
 namespace dataset {
 
 void AssertSchemasAre(std::vector<std::shared_ptr<Schema>> actual,
