+++ cpp/src/arrow/dataset/scanner_test.cc
@@ -30,7 +30,6 @@
 #include "arrow/compute/api_vector.h"
 #include "arrow/compute/cast.h"
 #include "arrow/compute/expression_internal.h"
-#include "arrow/compute/kernels/test_util_internal.h"
 #include "arrow/dataset/dataset_internal.h"
 #include "arrow/dataset/plan.h"
 #include "arrow/dataset/test_util_internal.h"
@@ -56,13 +55,8 @@ using testing::UnorderedElementsAreArray;
 
 namespace arrow {
 
-using compute::ComputeKernelEnvironment;
 using internal::GetCpuThreadPool;
 using internal::Iota;
-
-// Initialize the compute module
-::testing::Environment* compute_kernels_env =
-    ::testing::AddGlobalTestEnvironment(new ComputeKernelEnvironment);
 
 namespace dataset {
